Moved the queue_man function to ElevatorControl.c
The name was converted to queueManager, and it was changed as to use the
queue managing functions and the ElevatorData struct rather than the
linked list and node system.  The previous commit was preparing for this
shift over, by adding the funcitons needed to fully emulate the actions
of the linked list code.