Split ElevatorControl.c into many files
First, created a types file that contains the data for all the structs
and constants that are shared between the different files.  Then split
out the queue managing functions, and the functions that are going to be
managed in threads into one another.  The make file will compile
everything with a single command, and delete any auxillary object files
when it is done compiling.