added controlfunctions files and keylistener function
Beginning to prepare more to get everything in a full working order so
that all that needs to be added is the wiring-pi functions.  Split some
functions into a separate file so they can be included/moved out of
main.  Also added the function that will listen for key presses and pass
them into the queue manager.