Triggering door process working, and logging system
The doors will not close with regular interrupt, the doors open on a
reach floor interrupt.  Other misc changes.  Added a string-printing
logging system that makes it easy to turn on and off logging and change
the level at which logging is done without moving a lot of code around.