Bugfixes, plus testing should worked now
I made some small changes to the queue manager so that it would work better. More importantly, I figured out why the print queue wasn't working. It was called at the end of the function, but the function almost never actually ran it the end. Now I call it before every return and it seems to work. I didn't update the exe, because I wasn't sure if I was compiling it correctly, but it should compile fine.