Fixed Same-Floor Requests
If the Elevator was on the same floor as a request it wasn't doing
anything.  I assumed it was a bug with a queue manager, but all that
needed to be done was opening the doors.  The interrupt for key presses
now opens the doors and does not call queue manager if the elevator is
on the same floor: